Added simple runtime tracking
Additionally, did a little testing on implementations. As the comment in
freqTable says, I tried out using the standard python idiom for counting
something in a dictionary (d[key] = d.get(key, 0) + 1) since it's much
cleaner looking. Unfortunately, it turned out to be much slower. Loading
my system dictionary (~235,000 entries) it took ~3.7 seconds, compared
to 3.3 with the if-statements.