Added basic .txt file functionality
All this really does is read a text file like a dictionary (so,
stupidly). On a brighter note, I modularized things a bit better so that
there's a method for reading a file which is then called by both
process_dictionary and process_text. Just need to make txt reading
smarter now.